Move createUploadElement to its own file
