+++ src/ui/index.js
@@ -1,2 +1,3 @@
 export {default as addContainerDiv} from "./container";
+export {default as createUploadElement} from "./uploadElement";
 export {default as injectStyling} from "./styling";
